% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{airglms}
\alias{airglms}
\title{Select generalized linear models automatically}
\usage{
airglms(config_file, score, verbose = FALSE, sort = "default")
}
\arguments{
\item{config_file}{Name of the configuration file with analysis settings.}

\item{score}{Which information criteria to use (one of: AIC, BIC)}

\item{verbose}{Print intermediate results of GLM selection while running (default: FALSE)}

\item{sort}{How the dataframes in the resulting object will be sorted (default: in given order, score: by scores)}
}
\description{
The main function of the package.
It is used to automatically select generalized linear models (GLM).
The dependent and independent variables and interactions to be tested,
distributions and link functions, base model, data frame and
output file name are determined by the user in a configuration file.
GLM selection is performed with forward stepwise selection
using AIC or BIC.
Returns a character vector with models for each dependent variable,
and writes detailed output to the file specified.
}
\examples{
airglms("myconfig.txt", score="AIC")
}
