% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_interaction}
\alias{clean_interaction}
\title{Tidy interaction terms}
\usage{
clean_interaction(text_interaction)
}
\arguments{
\item{interaction}{A string representation of an interaction formula.}
}
\description{
This function is used to introduce consistent whitespaces in a text interaction formula.
Returns a character vector.
}
\examples{
extract_variables("a*b")
}
\keyword{interaction}
