% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_interaction_variables}
\alias{extract_interaction_variables}
\title{Extract interaction variables}
\usage{
extract_interaction_variables(interaction)
}
\arguments{
\item{interaction}{A string representation of an interaction formula.}
}
\description{
This function is used to extract variables from a text interaction formula.
Returns a list.
}
\examples{
extract_interaction_variables("a * b")
}
\keyword{variable}
