% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variables}
\usage{
extract_variables(text_formula)
}
\arguments{
\item{text_formula}{A string representation of a formula.}
}
\description{
This function is used to extract variables from a text formula.
Returns a list.
}
\examples{
extract_variables("x ~ a + b")
}
\keyword{variable}
