% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_parameter}
\alias{get_parameter}
\title{Get parameter}
\usage{
get_parameter(line, comma_split = NULL)
}
\arguments{
\item{line}{Line of text from the config file}

\item{comma_split}{Whether the line is split by commas. Default: no (NULL).}
}
\description{
This function is used to extract parameters from config file.
Returns text after equals sign.
}
\examples{
get_parameter("output = myfile.txt")
}
\keyword{parameter}
