% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_log}
\alias{write_log}
\title{Write to log file}
\usage{
write_log(file, ...)
}
\arguments{
\item{file}{Filename}

\item{...}{One or more R objects to be written as text to the file.}
}
\description{
This function is used to append a list of text to a file.
Does not return anything.
}
\examples{
write_log("mylog.txt", "Hello world!")
}
\keyword{log}
