% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{airglms}
\alias{airglms}
\title{Select generalized linear models automatically}
\usage{
airglms(config_file, verbose = FALSE)
}
\arguments{
\item{config_file}{Name of the configuration file with analysis settings.}

\item{verbose}{Print intermediate results of GLM selection while running (default: FALSE)}
}
\description{
The main function of the package.
It is used to automatically select generalized linear models (GLM).
The dependent and independent variables and interactions to be tested,
distributions and link functions, base model, data frame and
output file name are determined by the user in a configuration file.
GLM selection is performed with forward stepwise selection
using Akaike information criteria.
Returns a character vector with models for each dependent variable,
and writes detailed output to the file specified.
}
\examples{
airglms("myconfig.txt")
}
